+++ jd_plantBean.js
@@ -1,6 +1,7 @@
 /*
-种豆得豆 搬的https://github.com/uniqueque/QuantumultX/blob/4c1572d93d4d4f883f483f907120a75d925a693e/Script/jd_joy.js
-更新时间:2020-08-25
+种豆得豆 搬的https://github.com/uniqueque/QuantumultX/blob/4c1572d93d4d4f883f483f907120a75d925a693e/Script/jd_plantBean.js
+更新时间:2020-09-13
+已支持IOS京东双账号,云端N个京东账号
 脚本兼容: QuantumultX, Surge, Loon, JSBox, Node.js
 会自动关注任务中的店铺跟商品
 互助码shareCode请先手动运行脚本查看打印可看到
@@ -15,578 +16,614 @@ cron "1 7-21/2 * * *" script-path=https://raw.githubusercontent.com/lxk0301/scri
 一天只能帮助3个人。多出的助力码无效
 注：如果使用Node.js, 需自行安装'crypto-js,got,http-server,tough-cookie'模块. 例: npm install crypto-js http-server tough-cookie got --save
 */
-
-const name = '京东种豆得豆';
-const $ = new Env(name);
+const $ = new Env('京东种豆得豆');
 //Node.js用户请在jdCookie.js处填写京东ck;
-const jdCookieNode = $.isNode() ? require('./jdCookie.js') : '';
-
 //ios等软件用户直接用NobyDa的jd cookie
-const cookie = jdCookieNode.CookieJD ? jdCookieNode.CookieJD : $.getdata('CookieJD');
 let jdNotify = $.getdata('jdPlantBeanNotify');
-
+let cookiesArr = [], cookie = '', jdPlantBeanShareArr = [], isBox = false, notify, newShareCodes, option, message,subTitle;
 //京东接口地址
 const JD_API_HOST = 'https://api.m.jd.com/client.action';
-
-let plantUuids = [ // 这个列表填入你要助力的好友的plantUuid
-  '66j4yt3ebl5ierjljoszp7e4izzbzaqhi5k2unz2afwlyqsgnasq',
-  'olmijoxgmjutyrsovl2xalt2tbtfmg6sqldcb3q',
-  'qawf5ls3ucw25yhfulu32xekqy3h7wlwy7o5jii'
+//助力好友分享码(最多3个,否则后面的助力失败)
+//此此内容是IOS用户下载脚本到本地使用，填写互助码的地方，同一京东账号的好友互助码请使用@符号隔开。
+//下面给出两个账号的填写示例（iOS只支持2个京东账号）
+let shareCodes = [ // IOS本地脚本用户这个列表填入你要助力的好友的shareCode
+                   //账号一的好友shareCode,不同好友的shareCode中间用@符号隔开
+  '66j4yt3ebl5ierjljoszp7e4izzbzaqhi5k2unz2afwlyqsgnasq@olmijoxgmjutyrsovl2xalt2tbtfmg6sqldcb3q@e7lhibzb3zek27amgsvywffxx7hxgtzstrk2lba',
+  //账号二的好友shareCode,不同好友的shareCode中间用@符号隔开
+  '4npkonnsy7xi3p6pjfxg6ct5gll42gmvnz7zgoy@6dygkptofggtp6ffhbowku3xgu',
 ]
 let currentRoundId = null;//本期活动id
 let lastRoundId = null;//上期id
 let roundList = [];
 let awardState = '';//上期活动的京豆是否收取
-// 添加box功能
-// 【用box订阅的好处】
-// 1️⃣脚本也可以远程挂载了。助力功能只需在box里面设置助力码。
-// 2️⃣所有脚本的cookie都可以备份，方便你迁移到其他支持box的软件。
-let isBox = false //默认没有使用box
-const boxShareCodeArr = ['jd_plantBean1', 'jd_plantBean2', 'jd_plantBean3'];
-isBox = boxShareCodeArr.some((item) => {
-  const boxShareCode = $.getdata(item);
-  return (boxShareCode !== undefined && boxShareCode !== null && boxShareCode !== '');
-});
-if (isBox) {
-  plantUuids = [];
-  for (const item of boxShareCodeArr) {
-    if ($.getdata(item)) {
-      plantUuids.push($.getdata(item));
+
+!(async () => {
+  await requireConfig();
+  if (!cookiesArr[0]) {
+    $.msg($.name, '【提示】请先获取京东账号一cookie\n直接使用NobyDa的京东签到获取', 'https://bean.m.jd.com/', {"open-url": "https://bean.m.jd.com/"});
+    return;
+  }
+  for (let i = 0; i < cookiesArr.length; i++) {
+    if (cookiesArr[i]) {
+      cookie = cookiesArr[i];
+      UserName = decodeURIComponent(cookie.match(/pt_pin=(.+?);/) && cookie.match(/pt_pin=(.+?);/)[1])
+      $.index = i + 1;
+      console.log(`\n开始【京东账号${$.index}】${UserName}\n`);
+      message = '';
+      subTitle = '';
+      option = {};
+      await shareCodesFormat();
+      await jdPlantBean();
     }
   }
-}
+})().catch((e) => {
+  $.log('', `❌ ${$.name}, 失败! 原因: ${e}!`, '')
+}).finally(() => {
+  $.done();
+})
 
-var Task = step();
-Task.next();
-
-function* step() {
-  let message = '', subTitle = '';
-  if (cookie) {
-    console.log(`获取任务及基本信息`)
-    let plantBeanIndexResult = yield plantBeanIndex()
-    if (plantBeanIndexResult.code != "0") {
-      console.log(`plantBeanIndexResult:${JSON.stringify(plantBeanIndexResult)}`)
-      if (plantBeanIndexResult.code === '3') {
-        $.setdata('', 'CookieJD');//cookie失效，故清空cookie。
-        $.msg(name, '【提示】京东cookie已失效,请重新登录获取', 'https://bean.m.jd.com/', {"open-url": "https://bean.m.jd.com/"});
-        $.done();
-        return
-      }
-      //todo
-      return
-    }
-    roundList = plantBeanIndexResult.data.roundList;
-    currentRoundId = roundList[1].roundId;
-    lastRoundId = roundList[0].roundId;
+async function jdPlantBean() {
+  console.log(`获取任务及基本信息`)
+  await plantBeanIndex();
+  // console.log(plantBeanIndexResult.data.taskList);
+  if ($.plantBeanIndexResult.code === '0') {
+    const shareUrl = $.plantBeanIndexResult.data.jwordShareInfo.shareUrl
+    $.myPlantUuid = getParam(shareUrl, 'plantUuid')
+    console.log(`\n【您的互助码plantUuid】 ${$.myPlantUuid}\n`);
+    roundList = $.plantBeanIndexResult.data.roundList;
+    currentRoundId = roundList[1].roundId;//本期的roundId
+    lastRoundId = roundList[0].roundId;//上期的roundId
     awardState = roundList[0].awardState;
-    subTitle = `【京东昵称】${plantBeanIndexResult.data.plantUserInfo.plantNickName}`;
+    $.taskList = $.plantBeanIndexResult.data.taskList;
+    subTitle = `【京东昵称】${$.plantBeanIndexResult.data.plantUserInfo.plantNickName}`;
     message += `【上期时间】${roundList[0].dateDesc}\n`;
     message += `【上期成长值】${roundList[0].growth}\n`;
-    //定时领取--放到前面执行收取自动生产的营养液
-    if (plantBeanIndexResult.data.timeNutrientsRes.state == 1 && plantBeanIndexResult.data.timeNutrientsRes.nutrCount > 0) {
-      console.log(`开始领取定时产生的营养液`)
-      let receiveNutrientsResult = yield receiveNutrients(currentRoundId)
-      console.log(`receiveNutrientsResult:${JSON.stringify(receiveNutrientsResult)}`)
-    }
-    console.log(`【上轮京豆】${awardState === '4' ? '采摘中' : awardState === '5' ? '可收获了' : '已领取'}`);
-    if (awardState === '4') {
-      //京豆采摘中...
-      message += `【上期状态】${roundList[0].tipBeanEndTitle}\n`;
-    } else if (awardState === '5') {
-      //收获
-      let res = yield getReward();
-      // console.log(`种豆得豆收获的京豆情况---res,${JSON.stringify(res)}`);
-      console.log('开始领取京豆');
-      if (res.code === '0') {
-        console.log('京豆领取成功');
-        message += `【上期兑换京豆】${res.data.awardBean}个\n`;
-        $.msg(name, subTitle, message);
-      }
-    } else if (awardState === '6') {
-      //京豆已领取
-      message += `【上期兑换京豆】${roundList[0].awardBeans}个\n`;
-    }
-    if (roundList[1].dateDesc.indexOf('本期 ') > -1) {
-      roundList[1].dateDesc = roundList[1].dateDesc.substr(roundList[1].dateDesc.indexOf('本期 ') + 3, roundList[1].dateDesc.length);
-    }
-    message += `【本期时间】${roundList[1].dateDesc}\n`;
-    message += `【本期成长值】${roundList[1].growth}\n`;
-    let shareUrl = plantBeanIndexResult.data.jwordShareInfo.shareUrl
-    let myPlantUuid = getParam(shareUrl, 'plantUuid')
-    // console.log(`你的plantUuid为${myPlantUuid}`)
-    console.log(`\n【您的互助码plantUuid】 ${myPlantUuid}\n`);
-    for (let task of plantBeanIndexResult.data.taskList) {
-      console.log(`开始【${task.taskName}】任务`)
-      if (task.taskType == 7 || task.taskType == 17 || task.taskType == 18) {
-        //具体每个人可能不一样
-        //7金融双签,18疯抢爆品,17叠蛋糕
-        if (task.isFinished != 1) {
-          console.log(task.taskName)
-          let receiveNutrientsTaskResult = yield receiveNutrientsTask(task.taskType)
-          console.log(`receiveNutrientsTaskResult:${JSON.stringify(receiveNutrientsTaskResult)}`)
-        }
-      } else if (task.awardType == 3) {
-        //浏览店铺
-        if (task.isFinished != 1) {
-          let shopTaskListResult = yield shopTaskList()
-          if (shopTaskListResult.code == '0') {
-            let shops = shopTaskListResult.data.goodShopList.concat(shopTaskListResult.data.moreShopList)
-            let nutrCount = 0
-            for (let shop of shops) {
-              console.log(shop.shopName)
-              if (shop.taskState == '2') {
-                let shopNutrientsTaskResult = yield shopNutrientsTask(shop.shopTaskId, shop.shopId)
-                if (shopNutrientsTaskResult.code == 0) {
-                  if (shopNutrientsTaskResult.data.nutrState && shopNutrientsTaskResult.data.nutrState == '1' && shopNutrientsTaskResult.data.nutrCount > 0) {
-                    console.log(`关注店铺${shop.shopName}获得${shopNutrientsTaskResult.data.nutrCount}营养液`)
-                    nutrCount += shopNutrientsTaskResult.data.nutrCount
-                    if (nutrCount >= task.totalNum - task.gainedNum) {
-                      break
-                    }
-                  } else {
-                    console.log(`关注店铺${shop.shopName}未获得营养液`)
-                  }
-                } else {
-                  console.log(`${shop.shopName},shopNutrientsTaskResult:${JSON.stringify(shopNutrientsTaskResult)}`)
-                }
-              }
-            }
-          } else {
-            console.log(`shopTaskListResult:${JSON.stringify(shopTaskListResult)}`)
-          }
-        }
-      } else if (task.awardType == 10) {
-        //浏览频道
-        if (task.isFinished != 1) {
-          let plantChannelTaskListResult = yield plantChannelTaskList()
-          if (plantChannelTaskListResult.code == '0') {
-            let channelList = plantChannelTaskListResult.data.goodChannelList.concat(plantChannelTaskListResult.data.normalChannelList)
-            let nutrCount = 0
-            for (let channel of channelList) {
-              // console.log(channel.channelName)
-              if (channel.taskState == '2') {
-                let plantChannelNutrientsTaskResult = yield plantChannelNutrientsTask(channel.channelTaskId, channel.channelId)
-                if (plantChannelNutrientsTaskResult.code == '0') {
-                  if (plantChannelNutrientsTaskResult.data.nutrState && plantChannelNutrientsTaskResult.data.nutrState == '1' && plantChannelNutrientsTaskResult.data.nutrNum > 0) {
-                    console.log(`浏览频道${channel.channelName}获得${plantChannelNutrientsTaskResult.data.nutrNum}营养液`)
-                    nutrCount += plantChannelNutrientsTaskResult.data.nutrNum
-                    if (nutrCount >= task.totalNum - task.gainedNum) {
-                      break
-                    }
-                  } else {
-                    console.log(`浏览频道${channel.channelName}未获得营养液`)
-                  }
-                } else {
-                  console.log(`${channel.channelName},plantChannelNutrientsTaskResult:${JSON.stringify(plantChannelNutrientsTaskResult)}`)
-
-                }
-              }
-            }
-          } else {
-            console.log(`plantChannelTaskListResult:${JSON.stringify(plantChannelTaskListResult)}`)
-          }
-        }
-      } else if (task.awardType == 5) {
-        //关注商品
-        if (task.isFinished != 1) {
-          let productTaskListResult = yield productTaskList()
-          if (productTaskListResult.code == '0') {
-            let productInfoList = productTaskListResult.data.productInfoList.map(([item]) => item)
-            let nutrCount = 0
-            for (let productInfo of productInfoList) {
-              console.log(productInfo.productName)
-              if (productInfo.taskState == '2') {
-                let productNutrientsTaskResult = yield productNutrientsTask(productInfo.productTaskId, productInfo.skuId)
-                if (productNutrientsTaskResult.code == '0') {
-                  if (productNutrientsTaskResult.data.nutrState && productNutrientsTaskResult.data.nutrState == '1' && productNutrientsTaskResult.data.nutrCount > 0) {
-                    console.log(`关注商品${productInfo.productName}获得${productNutrientsTaskResult.data.nutrCount}营养液`)
-                    nutrCount += productNutrientsTaskResult.data.nutrCount
-                    if (nutrCount >= task.totalNum - task.gainedNum) {
-                      break
-                    }
-                  } else {
-                    console.log(`关注商品${productInfo.productName}未获得营养液`)
-                  }
-                } else {
-                  console.log(`productNutrientsTaskResult:${JSON.stringify(productNutrientsTaskResult)}`)
-                }
-              }
-            }
-          } else {
-            console.log(`productTaskListResult:${JSON.stringify(productTaskListResult)}`)
-          }
-        }
-      } else if (task.taskType == 4) {
-        //逛逛会场
-        if (task.isFinished != 1 && task.gainedNum == '0') {
-          if (plantBeanIndexResult.data.roundList[1].roundState == 2) {
-            let purchaseRewardTaskResult = yield purchaseRewardTask(plantBeanIndexResult.data.roundList[1].roundId)
-            console.log(`purchaseRewardTaskResult:${JSON.stringify(purchaseRewardTaskResult)}`)
-          }
-        }
-      } else if (task.taskType == 19) {
-        // 低价包邮
-        if (task.isFinished !== 1) {
-          let plantReceiveNutrientsTaskRes = yield plantReceiveNutrientsTask();
-          console.log(`${task.taskName}获取营养液：：${plantReceiveNutrientsTaskRes.data && plantReceiveNutrientsTaskRes.data.nutrNum}`)
-        }
-      } else if (task.taskType == 20) {
-        // 助力高考
-        if (task.isFinished !== 1) {
-          let plantReceiveNutrientsTaskRes = yield receiveNutrientsTask(task.taskType);
-          console.log(`${task.taskName}获取营养液：：${plantReceiveNutrientsTaskRes.data && plantReceiveNutrientsTaskRes.data.nutrNum}`)
-        }
-      } else if (task.taskType == 1) {
-        console.log('跳过签到，NobyDa的会签')
-        // console.log(`【${task.taskName}】未开发${task.awardType},${task.taskType}`)
-      } else {
-        console.log(`【${task.taskName}】未开发${task.awardType},${task.taskType}`)
-      }
-      console.log(`【${task.taskName}】任务结束`)
-    }
-
-    //任务列表少了金融双签，拉出来执行下
-    console.log(`金融双签`)
-    let receiveNutrientsTaskResult = yield receiveNutrientsTask(7)
-    console.log(`receiveNutrientsTaskResult:${JSON.stringify(receiveNutrientsTaskResult)}`)
-
-    //助力好友
-    console.log('开始助力好友')
-    for (let plantUuid of plantUuids) {
-      if (plantUuid == myPlantUuid) {
-        console.log('跳过自己的plantUuid')
-        continue
-      }
-      console.log(`开始助力好友: ${plantUuid}`);
-      let helpResult = yield helpShare(plantUuid)
-      if (helpResult.code === '0') {
-        console.log(`助力好友结果: ${JSON.stringify(helpResult.data.helpShareRes)}`);
-        if (helpResult.data.helpShareRes && helpResult.data.helpShareRes.state === '2') {
-          console.log('今日助力机会已耗尽，跳出助力');
-          break;
-        }
-      } else {
-        console.log(`助力好友失败: ${JSON.stringify(helpResult)}`);
-      }
-    }
-
-    //天天扭蛋功能
-    let eggChance = yield egg();
-    if (eggChance.code == 0) {
-      if (eggChance.data.restLotteryNum > 0) {
-        const eggL = new Array(eggChance.data.restLotteryNum).fill('');
-        for (let i = 0; i < eggL.length; i++) {
-          console.log(`开始第${i + 1}次扭蛋`);
-          let plantEggDoLotteryRes = yield plantEggDoLottery();
-          console.log(`天天扭蛋成功：${JSON.stringify(plantEggDoLotteryRes)}`);
-        }
-      } else {
-        console.log('暂无扭蛋机会')
-      }
-    } else {
-      console.log('查询天天扭蛋的机会失败')
-    }
-    plantBeanIndexResult = yield plantBeanIndex()
-    if (plantBeanIndexResult.code == '0') {
-      let plantBeanRound = plantBeanIndexResult.data.roundList[1]
-      if (plantBeanRound.roundState == 2) {
-        //收取营养液
-        console.log(`开始收取营养液`)
-        for (let bubbleInfo of plantBeanRound.bubbleInfos) {
-          console.log(`收取营养液${bubbleInfo.name}`)
-          let cultureBeanResult = yield cultureBean(plantBeanRound.roundId, bubbleInfo.nutrientsType)
-          console.log(`cultureBeanResult:${JSON.stringify(cultureBeanResult)}`)
-        }
-      }
-    } else {
-      console.log(`plantBeanIndexResult:${JSON.stringify(plantBeanIndexResult)}`)
-    }
-    // 偷大于等于3瓶好友的营养液
-    let stealRes = yield steal();
-    if (stealRes.code == 0) {
-      if (stealRes.data.tips) {
-        console.log('今日已达上限');
-      }
-      if (stealRes.data && stealRes.data.friendInfoList && stealRes.data.friendInfoList.length > 0) {
-        for (let item of stealRes.data.friendInfoList) {
-          if (item.nutrCount >= 3) {
-            console.log(`可以偷的好友的信息::${JSON.stringify(item)}`);
-            console.log(`可以偷的好友的信息paradiseUuid::${JSON.stringify(item.paradiseUuid)}`);
-            let stealFriendRes = yield collectUserNutr(item.paradiseUuid);
-            console.log(`偷取好友营养液情况:${JSON.stringify(stealFriendRes)}`)
-            if (stealFriendRes.code == '0') {
-              console.log(`偷取好友营养液成功`)
-            }
-          }
-        }
-      }
-    }
-    console.log('结束')
+    await receiveNutrients();//定时领取营养液
+    await doHelp();//助力
+    await doTask();//做日常任务
+    await doEgg();
+    await stealFriendWater();
+    await doCultureBean();
+    await doGetReward();
+    await showTaskProcess();
+    await showMsg();
   } else {
-    $.msg(name, '【提示】请先获取cookie\n直接使用NobyDa的京东签到获取', 'https://bean.m.jd.com/', {"open-url": "https://bean.m.jd.com/"});
-    $.done();
-    return
+    if ($.plantBeanIndexResult.code === '3') {
+      //cookie过期
+      $.msg($.name, '【提示】京东cookie已失效,请重新登录获取', 'https://bean.m.jd.com/', {"open-url": "https://bean.m.jd.com/"});
+      if ($.index === 1) {
+        $.setdata('', 'CookieJD');//cookie失效，故清空cookie。
+      } else if ($.index === 2){
+        $.setdata('', 'CookieJD2');//cookie失效，故清空cookie。
+      }
+      if ($.isNode()) {
+        await notify.sendNotify(`${$.name}cookie已失效`, `京东账号${$.index} ${UserName}\n\n请重新登录获取cookie`);
+      }
+      if ($.isNode()) {
+        await notify.BarkNotify(`${$.name}cookie已失效`, `京东账号${$.index} ${UserName}\n请重新登录获取cookie`);
+      }
+    } else {
+      console.log(`种豆得豆-初始失败:  ${JSON.stringify($.plantBeanIndexResult)}`);
+    }
   }
+}
+async function doGetReward() {
+  console.log(`【上轮京豆】${awardState === '4' ? '采摘中' : awardState === '5' ? '可收获了' : '已领取'}`);
+  if (awardState === '4') {
+    //京豆采摘中...
+    message += `【上期状态】${roundList[0].tipBeanEndTitle}\n`;
+  } else if (awardState === '5') {
+    //收获
+    await getReward();
+    console.log('开始领取京豆');
+    if ($.getReward.code === '0') {
+      console.log('京豆领取成功');
+      message += `【上期兑换京豆】${$.getReward.data.awardBean}个\n`;
+      $.msg(name, subTitle, message);
+      if ($.isNode()) {
+        const notifyMessage = message.replace(/[\n\r]/g, '\n\n');
+        await notify.sendNotify(`${$.name}`, `京东账号${$.index} ${UserName}\n\n${notifyMessage}`);
+      }
+      if ($.isNode()) {
+        await notify.BarkNotify(`${$.name}`, `京东账号${$.index} ${UserName}\n${message}`);
+      }
+    }
+  } else if (awardState === '6') {
+    //京豆已领取
+    message += `【上期兑换京豆】${roundList[0].awardBeans}个\n`;
+  }
+  if (roundList[1].dateDesc.indexOf('本期 ') > -1) {
+    roundList[1].dateDesc = roundList[1].dateDesc.substr(roundList[1].dateDesc.indexOf('本期 ') + 3, roundList[1].dateDesc.length);
+  }
+  message += `【本期时间】${roundList[1].dateDesc}\n`;
+  message += `【本期成长值】${roundList[1].growth}\n`;
+}
+async function doCultureBean() {
+  await plantBeanIndex();
+  if ($.plantBeanIndexResult.code === '0') {
+    const plantBeanRound = $.plantBeanIndexResult.data.roundList[1]
+    if (plantBeanRound.roundState === '2') {
+      //收取营养液
+      console.log(`开始收取营养液`)
+      for (let bubbleInfo of plantBeanRound.bubbleInfos) {
+        console.log(`收取-${bubbleInfo.name}-的营养液`)
+        await cultureBean(plantBeanRound.roundId, bubbleInfo.nutrientsType)
+        console.log(`收取营养液结果:${JSON.stringify($.cultureBeanRes)}`)
+      }
+    }
+  } else {
+    console.log(`plantBeanIndexResult:${JSON.stringify($.plantBeanIndexResult)}`)
+  }
+}
+async function stealFriendWater() {
+  await stealFriendList();
+  if ($.stealFriendList.code === '0') {
+    if ($.stealFriendList.data.tips) {
+      console.log('偷取好友营养液今日已达上限');
+      return
+    }
+    if ($.stealFriendList.data && $.stealFriendList.data.friendInfoList && $.stealFriendList.data.friendInfoList.length > 0) {
+      for (let item of $.stealFriendList.data.friendInfoList) {
+        if (item.nutrCount >= 3) {
+          // console.log(`可以偷的好友的信息::${JSON.stringify(item)}`);
+          console.log(`可以偷的好友的信息paradiseUuid::${JSON.stringify(item.paradiseUuid)}`);
+          await collectUserNutr(item.paradiseUuid);
+          console.log(`偷取好友营养液情况:${JSON.stringify($.stealFriendRes)}`)
+          if ($.stealFriendRes.code === '0') {
+            console.log(`偷取好友营养液成功`)
+          }
+        }
+      }
+    }
+  }
+}
+async function doEgg() {
+  await egg();
+  if ($.plantEggLotteryRes.code === '0') {
+    if ($.plantEggLotteryRes.data.restLotteryNum > 0) {
+      const eggL = new Array($.plantEggLotteryRes.data.restLotteryNum).fill('');
+      console.log(`目前共有${eggL.length}次扭蛋的机会`)
+      for (let i = 0; i < eggL.length; i++) {
+        console.log(`开始第${i + 1}次扭蛋`);
+        await plantEggDoLottery();
+        console.log(`天天扭蛋成功：${JSON.stringify($.plantEggDoLotteryResult)}`);
+      }
+    } else {
+      console.log('暂无扭蛋机会')
+    }
+  } else {
+    console.log('查询天天扭蛋的机会失败')
+  }
+}
+async function doTask() {
+  if ($.taskList && $.taskList.length > 0) {
+    for (let item of $.taskList) {
+      if (item.isFinished === 1) {
+        console.log(`${item.taskName} 任务已完成\n`);
+        continue;
+      }
+      if (item.dailyTimes === 1 && item.taskType !== 8) {
+        console.log(`\n开始做 ${item.taskName}任务`);
+        // $.receiveNutrientsTaskRes = await receiveNutrientsTask(item.taskType);
+        await receiveNutrientsTask(item.taskType);
+        console.log(`做 ${item.taskName}任务结果:${JSON.stringify($.receiveNutrientsTaskRes)}\n`);
+      }
+      if (item.taskType === 3) {
+        //浏览店铺
+        console.log(`开始做 ${item.taskName}任务`);
+        let unFinishedShopNum = item.totalNum - item.gainedNum;
+        if (unFinishedShopNum === 0) {
+          continue
+        }
+        await shopTaskList();
+        const { data } = $.shopTaskListRes;
+        let goodShopListARR = [], moreShopListARR = [], shopList = [];
+        const { goodShopList, moreShopList } = data;
+        for (let i of goodShopList) {
+          if (i.taskState === '2') {
+            goodShopListARR.push(i);
+          }
+        }
+        for (let j of moreShopList) {
+          if (j.taskState === '2') {
+            moreShopListARR.push(j);
+          }
+        }
+        shopList = goodShopListARR.concat(moreShopListARR);
+        for (let shop of shopList) {
+          const { shopId, shopTaskId } = shop;
+          const body = {
+            "monitor_refer": "plant_shopNutrientsTask",
+            "shopId": shopId,
+            "shopTaskId": shopTaskId
+          }
+          const shopRes = await requestGet('shopNutrientsTask', body);
+          console.log(`shopRes结果:${JSON.stringify(shopRes)}`);
+          if (shopRes.code === '0') {
+            if (shopRes.data.nutrState === '1') {
+              unFinishedShopNum --;
+            }
+          }
+          if (unFinishedShopNum <= 0) {
+            console.log(`${item.taskName}任务已做完\n`)
+            break;
+          }
+        }
+      }
+      if (item.taskType === 5) {
+        //挑选商品
+        console.log(`开始做 ${item.taskName}任务`);
+        let unFinishedProductNum = item.totalNum - item.gainedNum;
+        if (unFinishedProductNum === 0) {
+          continue
+        }
+        await productTaskList();
+        // console.log('productTaskList', $.productTaskList);
+        const { data } = $.productTaskList;
+        let productListARR = [], productList = [];
+        const { productInfoList } = data;
+        for (let i = 0; i < productInfoList.length; i++) {
+          for (let j = 0; j < productInfoList[i].length; j++){
+            productListARR.push(productInfoList[i][j]);
+          }
+        }
+        for (let i of productListARR) {
+          if (i.taskState === '2') {
+            productList.push(i);
+          }
+        }
+        for (let product of productList) {
+          const { skuId, productTaskId } = product;
+          const body = {
+            "monitor_refer": "plant_productNutrientsTask",
+            "productTaskId": productTaskId,
+            "skuId": skuId
+          }
+          const productRes = await requestGet('productNutrientsTask', body);
+          if (productRes.code === '0') {
+            // console.log('nutrState', productRes)
+            //这里添加多重判断,有时候会出现活动太火爆的问题,导致nutrState没有
+            if (productRes.data && productRes.data.nutrState && productRes.data.nutrState === '1') {
+              unFinishedProductNum --;
+            }
+          }
+          if (unFinishedProductNum <= 0) {
+            console.log(`${item.taskName}任务已做完\n`)
+            break;
+          }
+        }
+      }
+      if (item.taskType === 10) {
+        //关注频道
+        console.log(`开始做 ${item.taskName}任务`);
+        let unFinishedChannelNum = item.totalNum - item.gainedNum;
+        if (unFinishedChannelNum === 0) {
+          continue
+        }
+        await plantChannelTaskList();
+        const { data } = $.plantChannelTaskList;
+        // console.log('goodShopList', data.goodShopList);
+        // console.log('moreShopList', data.moreShopList);
+        let goodChannelListARR = [], normalChannelListARR = [], channelList = [];
+        const { goodChannelList, normalChannelList } = data;
+        for (let i of goodChannelList) {
+          if (i.taskState === '2') {
+            goodChannelListARR.push(i);
+          }
+        }
+        for (let j of normalChannelList) {
+          if (j.taskState === '2') {
+            normalChannelListARR.push(j);
+          }
+        }
+        channelList = goodChannelListARR.concat(normalChannelListARR);
+        for (let channelItem of channelList) {
+          const { channelId, channelTaskId } = channelItem;
+          const body = {
+            "channelId": channelId,
+            "channelTaskId": channelTaskId
+          }
+          const channelRes = await requestGet('plantChannelNutrientsTask', body);
+          console.log(`channelRes结果:${JSON.stringify(channelRes)}`);
+          if (channelRes.code === '0') {
+            if (channelRes.data.nutrState === '1') {
+              unFinishedChannelNum --;
+            }
+          }
+          if (unFinishedChannelNum <= 0) {
+            console.log(`${item.taskName}任务已做完\n`)
+            break;
+          }
+        }
+      }
+    }
+  }
+}
+function showTaskProcess() {
+  return new Promise(async resolve => {
+    await plantBeanIndex();
+    $.taskList = $.plantBeanIndexResult.data.taskList;
+    if ($.taskList && $.taskList.length > 0) {
+      console.log("     任务   进度");
+      for (let item of $.taskList) {
+        console.log(`[${item["taskName"]}]  ${item["gainedNum"]}/${item["totalNum"]}   ${item["isFinished"]}`);
+      }
+    }
+    resolve()
+  })
+}
+//助力好友
+async function doHelp() {
+  for (let plantUuid of newShareCodes) {
+    if (plantUuid === $.myPlantUuid) {
+      console.log(`\n跳过自己的plantUuid\n`)
+      continue
+    }
+    console.log(`\n开始助力好友: ${plantUuid}`);
+    await helpShare(plantUuid);
+    if ($.helpResult.code === '0') {
+      // console.log(`助力好友结果: ${JSON.stringify($.helpResult.data.helpShareRes)}`);
+      if ($.helpResult.data.helpShareRes) {
+        if ($.helpResult.data.helpShareRes.state === '1') {
+          console.log(`助力好友${plantUuid}成功`)
+          console.log(`${$.helpResult.data.helpShareRes.promptText}\n`);
+        } else if ($.helpResult.data.helpShareRes.state === '2') {
+          console.log('您今日助力的机会已耗尽，已不能再帮助好友助力了\n');
+          break;
+        } else if ($.helpResult.data.helpShareRes.state === '3') {
+          console.log('该好友今日已满20人助力,明天再来为Ta助力吧\n')
+        } else if ($.helpResult.data.helpShareRes.state === '4') {
+          console.log(`${$.helpResult.data.helpShareRes.promptText}\n`)
+        }
+      }
+    } else {
+      console.log(`助力好友失败: ${JSON.stringify($.helpResult)}`);
+    }
+  }
+}
+function showMsg() {
   if (!jdNotify || jdNotify === 'false') {
-    $.msg(name, subTitle, message);
+    $.msg($.name, subTitle, message);
   }
-  $.done();
 }
-
-function purchaseRewardTask(roundId) {
-  let functionId = arguments.callee.name.toString();
-  let body = {
-    "monitor_refer": "plant_receiveNutrients",
-    "monitor_source": "plant_app_plant_index",
-    "roundId": roundId,
-    "version": "9.0.0.1"
-  }
-  request(functionId, body);// `body=${escape(JSON.stringify(body))}&uuid=&appid=ld`
-}
-
-//低价包邮
-function plantReceiveNutrientsTask() {
+// ================================================此处是API=================================
+//每轮种豆活动获取结束后,自动收取京豆
+async function getReward() {
   const body = {
-    "monitor_refer": "plant_receiveNutrientsTask",
-    "monitor_source": "plant_app_plant_index",
-    "awardType": "19",
-    "version": "9.0.0.1"
-  };
-  request('receiveNutrientsTask', body);
-}
-
-function receiveNutrientsTask(awardType) {
-  // let functionId = arguments.callee.name.toString();
-  // let body = {
-  //     "monitor_refer": "plant_receiveNutrientsTask",
-  //     "monitor_source": "plant_m_plant_index",//plant_app_plant_index,plant_m_plant_index
-  //     "awardType": `"${awardType}"`,
-  //     "version": "9.0.0.1"// "9.0.0.1", "8.4.0.0"
-  // }
-  //这里很奇怪，试了很多情况都不行，直接这样了
-  requestGet(`https://api.m.jd.com/client.action?functionId=receiveNutrientsTask&body=%7B%22awardType%22%3A%22${awardType}%22%2C%22monitor_source%22%3A%22plant_m_plant_index%22%2C%22monitor_refer%22%3A%22plant_receiveNutrientsTask%22%2C%22version%22%3A%228.4.0.0%22%7D&appid=ld&client=apple&clientVersion=&networkType=&osVersion=&uuid=`)
-  // request(functionId, body);// `body=${escape(JSON.stringify(body))}&client=apple&appid=ld`
-}
-
-//https://api.m.jd.com/client.action?functionId=receiveNutrients
-function receiveNutrients(roundId) {
-
-  let functionId = arguments.callee.name.toString();
-  let body = {
-    "monitor_refer": "plant_receiveNutrients",
-    "monitor_source": "plant_app_plant_index",
-    "roundId": roundId,
-    "version": "9.0.0.1"
+    "roundId": lastRoundId
   }
-
-  request(functionId, body);//`body=${escape(JSON.stringify(body))}&uuid=&appid=ld`
-
+  $.getReward = await request('receivedBean', body);
 }
-
-// https://api.m.jd.com/client.action?functionId=cultureBean
 //收取营养液
-function cultureBean(roundId, nutrientsType) {
+async function cultureBean(currentRoundId, nutrientsType) {
   let functionId = arguments.callee.name.toString();
   let body = {
-    "monitor_refer": "plant_index",
-    "monitor_source": "plant_app_plant_index",
-    "roundId": roundId,
+    "roundId": currentRoundId,
     "nutrientsType": nutrientsType,
-    "version": "9.0.0.1"
   }
-  request(functionId, body);//`body=${escape(JSON.stringify(body))}&uuid=&appid=ld`
+  $.cultureBeanRes = await request(functionId, body);
 }
-
-function productNutrientsTask(productTaskId, skuId) {
-  let functionId = arguments.callee.name.toString();
-  let body = {
-    "monitor_refer": "plant_productNutrientsTask",
-    "monitor_source": "plant_app_plant_index",
-    "productTaskId": productTaskId,
-    "skuId": skuId,
-    "version": "9.0.0.1"
-  }
-  request(functionId, body);//`body=${escape(JSON.stringify(body))}&uuid=&appid=ld`
-}
-
-function productTaskList() {
-  //https://api.m.jd.com/client.action?functionId=productTaskList&body=%7B%7D&uuid=&appid=ld
-  let functionId = arguments.callee.name.toString();
-  request(functionId);// `body=%7B%7D&uuid=&appid=ld`
-}
-
-function plantChannelNutrientsTask(channelTaskId, channelId) {
-  let functionId = arguments.callee.name.toString();
-  let body = {"channelTaskId": channelTaskId, "channelId": channelId}
-  request(functionId, body);//`body=${escape(JSON.stringify(body))}&uuid=&appid=ld`
-}
-
-function plantChannelTaskList() {
-  let functionId = arguments.callee.name.toString();
-  request(functionId);// `body=%7B%7D&uuid=&appid=ld`
-}
-
-function shopNutrientsTask(shopTaskId, shopId) {
-  let functionId = arguments.callee.name.toString();
-  let body = {
-    "version": "9.0.0.1",
-    "monitor_refer": "plant_shopNutrientsTask",
-    "monitor_source": "plant_app_plant_index",
-    "shopId": shopId,
-    "shopTaskId": shopTaskId
-  }
-
-  request(functionId, body);// `body=${escape(JSON.stringify(body))}&uuid=&appid=ld`
-}
-
-function shopTaskList() {
-  let functionId = arguments.callee.name.toString();
-  request(functionId);//`body=%7B%7D&uuid=&appid=ld`
-}
-
-function helpShare(plantUuid) {
-  let body = {
-    "plantUuid": plantUuid,
-    "monitor_refer": "",
-    "wxHeadImgUrl": "",
-    "shareUuid": "",
-    "followType": "0",
-    "monitor_source": "plant_m_plant_index",
-    "version": "9.0.0.1"
-  }
-  request(`plantBeanIndex`, body);
-}
-
-//查询天天扭蛋的机会
-function egg() {
-  request('plantEggLotteryIndex');
-}
-
-// 调用扭蛋api
-function plantEggDoLottery() {
-  request('plantEggDoLottery');
-}
-
-function plantBeanIndex() {
-  // https://api.m.jd.com/client.action?functionId=plantBeanIndex
-  let functionId = arguments.callee.name.toString();
-  let body = {"monitor_source": "plant_app_plant_index", "monitor_refer": "", "version": "9.0.0.1"}
-  request(functionId, body);//plantBeanIndexBody
-}
-
 //偷营养液大于等于3瓶的好友
 //①查询好友列表
-function steal() {
+async function stealFriendList() {
   const body = {
     pageNum: '1'
   }
-  request('plantFriendList', body);
+  $.stealFriendList = await request('plantFriendList', body);
 }
 
 //②执行偷好友营养液的动作
-function collectUserNutr(paradiseUuid) {
+async function collectUserNutr(paradiseUuid) {
   console.log('开始偷好友');
-  console.log(paradiseUuid);
+  // console.log(paradiseUuid);
   let functionId = arguments.callee.name.toString();
   const body = {
     "paradiseUuid": paradiseUuid,
     "roundId": currentRoundId
   }
-  request(functionId, body);
+  $.stealFriendRes = await request(functionId, body);
 }
-
-//每轮种豆活动获取结束后,自动收取京豆
-function getReward() {
+async function receiveNutrients() {
+  $.receiveNutrientsRes = await request('receiveNutrients', {"roundId": currentRoundId, "monitor_refer": "plant_receiveNutrients"})
+  // console.log(`定时领取营养液结果:${JSON.stringify($.receiveNutrientsRes)}`)
+}
+async function plantEggDoLottery() {
+  $.plantEggDoLotteryResult = await requestGet('plantEggDoLottery');
+}
+//查询天天扭蛋的机会
+async function egg() {
+  $.plantEggLotteryRes = await requestGet('plantEggLotteryIndex');
+}
+async function productTaskList() {
+  let functionId = arguments.callee.name.toString();
+  $.productTaskList = await requestGet(functionId, {"monitor_refer": "plant_productTaskList"});
+}
+async function plantChannelTaskList() {
+  let functionId = arguments.callee.name.toString();
+  $.plantChannelTaskList = await requestGet(functionId);
+  // console.log('$.plantChannelTaskList', $.plantChannelTaskList)
+}
+async function shopTaskList() {
+  let functionId = arguments.callee.name.toString();
+  $.shopTaskListRes = await requestGet(functionId, {"monitor_refer": "plant_receiveNutrients"});
+  // console.log('$.shopTaskListRes', $.shopTaskListRes)
+}
+async function receiveNutrientsTask(awardType) {
+  const functionId = arguments.callee.name.toString();
   const body = {
-    "roundId": lastRoundId
+    "monitor_refer": "receiveNutrientsTask",
+    "awardType": `${awardType}`,
   }
-  request('receivedBean', body);
+  $.receiveNutrientsTaskRes = await requestGet(functionId, body);
 }
-
-function requestGet(url) {
-  const option = {
-    url: url,
-    headers: {
-      Cookie: cookie,
+//助力好友的api
+async function helpShare(plantUuid) {
+  const body = {
+    "plantUuid": plantUuid,
+    "wxHeadImgUrl": "",
+    "shareUuid": "",
+    "followType": "1",
+  }
+  $.helpResult = await request(`plantBeanIndex`, body);
+}
+async function plantBeanIndex() {
+  $.plantBeanIndexResult = await request('plantBeanIndex');//plantBeanIndexBody
+}
+//格式化助力码
+function shareCodesFormat() {
+  return new Promise(resolve => {
+    console.log(`第${$.index}个京东账号的助力码:::${jdPlantBeanShareArr[$.index - 1]}`)
+    if (jdPlantBeanShareArr[$.index - 1]) {
+      newShareCodes = jdPlantBeanShareArr[$.index - 1].split('@');
+    } else {
+      console.log(`由于您未提供shareCode,将采纳本脚本自带的助力码\n`)
+      const tempIndex = $.index > shareCodes.length ? (shareCodes.length - 1) : ($.index - 1);
+      newShareCodes = shareCodes[tempIndex].split('@');
     }
-  };
-  $.get(option, (err, resp, data) => {
-    try {
-      if (err) {
-        console.log('\n种豆得豆: API查询请求失败 ‼️‼️')
-      } else {
-        data = JSON.parse(data);
-      }
-    } catch (e) {
-      $.logErr(e, resp)
-    } finally {
-      sleep(data);
-    }
-    // try {
-    //   sleep(JSON.parse(data))
-    // } catch (e) {
-    //   $.logErr(e, resp)
-    // }
+    console.log(`格式化后第${$.index}个京东账号的助力码${JSON.stringify(newShareCodes)}`)
+    resolve();
   })
 }
-
-function request(function_id, body = {}) {
-  $.post(taskurl(function_id, body), (err, resp, data) => {
-    try {
-      if (err) {
-        console.log('\n种豆得豆: API查询请求失败 ‼️‼️')
-      } else {
-        data = JSON.parse(data);
-      }
-    } catch (e) {
-      $.logErr(e, resp)
-    } finally {
-      sleep(data);
+function requireConfig() {
+  return new Promise(resolve => {
+    console.log('开始获取种豆得豆配置文件\n')
+    notify = $.isNode() ? require('./sendNotify') : '';
+    //Node.js用户请在jdCookie.js处填写京东ck;
+    const jdCookieNode = $.isNode() ? require('./jdCookie.js') : '';
+    const jdPlantBeanShareCodes = $.isNode() ? require('./jdPlantBeanShareCodes.js') : '';
+    //IOS等用户直接用NobyDa的jd cookie
+    if ($.isNode()) {
+      Object.keys(jdCookieNode).forEach((item) => {
+        if (jdCookieNode[item]) {
+          cookiesArr.push(jdCookieNode[item])
+        }
+      })
+    } else {
+      cookiesArr.push($.getdata('CookieJD'));
+      cookiesArr.push($.getdata('CookieJD2'));
     }
+    console.log(`共${cookiesArr.length}个京东账号\n`)
+    if ($.isNode()) {
+      Object.keys(jdPlantBeanShareCodes).forEach((item) => {
+        if (jdPlantBeanShareCodes[item]) {
+          jdPlantBeanShareArr.push(jdPlantBeanShareCodes[item])
+        }
+      })
+    } else {
+      const boxShareCodeArr = ['jd_plantBean1', 'jd_plantBean2', 'jd_plantBean3'];
+      const boxShareCodeArr2 = ['jd2_plantBean1', 'jd2_plantBean2', 'jd2_plantBean3'];
+      const isBox1 = boxShareCodeArr.some((item) => {
+        const boxShareCode = $.getdata(item);
+        return (boxShareCode !== undefined && boxShareCode !== null && boxShareCode !== '');
+      });
+      const isBox2 = boxShareCodeArr2.some((item) => {
+        const boxShareCode = $.getdata(item);
+        return (boxShareCode !== undefined && boxShareCode !== null && boxShareCode !== '');
+      });
+      isBox = isBox1 ? isBox1 : isBox2;
+      if (isBox1) {
+        let temp = [];
+        for (const item of boxShareCodeArr) {
+          if ($.getdata(item)) {
+            temp.push($.getdata(item))
+          }
+        }
+        jdPlantBeanShareArr.push(temp.join('@'));
+      }
+      if (isBox2) {
+        let temp = [];
+        for (const item of boxShareCodeArr2) {
+          if ($.getdata(item)) {
+            temp.push($.getdata(item))
+          }
+        }
+        jdPlantBeanShareArr.push(temp.join('@'));
+      }
+    }
+    console.log(`\n种豆得豆助力码::${JSON.stringify(jdPlantBeanShareArr)}`);
+    console.log(`您提供了${jdPlantBeanShareArr.length}个账号的种豆得豆助力码\n`);
+    resolve()
   })
 }
-
-function taskurl(function_id, body) {
-  // console.log(`${JD_API_HOST}?functionId=${function_id}&body=${escape(JSON.stringify(body))}&appid=ld&client=apple&clientVersion=&networkType=&osVersion=&uuid=`)
+function requestGet(function_id, body = {}) {
+  if (!body.version) {
+    body["version"] = "9.0.0.1";
+  }
+  body["monitor_source"] = "plant_app_plant_index";
+  body["monitor_refer"] = "";
+  return new Promise(async resolve => {
+    await $.wait(2000);
+    const option = {
+      url: `${JD_API_HOST}?functionId=${function_id}&body=${escape(JSON.stringify(body))}&appid=ld`,
+      headers: {
+        'Cookie': cookie,
+        'Host': 'api.m.jd.com',
+        'Accept': '*/*',
+        'Connection': 'keep-alive',
+        'User-Agent': 'JD4iPhone/167283 (iPhone;iOS 13.6.1;Scale/3.00)',
+        'Accept-Language': 'zh-Hans-CN;q=1,en-CN;q=0.9',
+        'Accept-Encoding': 'gzip, deflate, br',
+        'Content-Type': "application/x-www-form-urlencoded"
+      }
+    };
+    $.get(option, (err, resp, data) => {
+      try {
+        if (err) {
+          console.log('\n种豆得豆: API查询请求失败 ‼️‼️')
+          $.logErr(err);
+        } else {
+          data = JSON.parse(data);
+        }
+      } catch (e) {
+        $.logErr(e, resp);
+      } finally {
+        resolve(data);
+      }
+    })
+  })
+}
+function request(function_id, body = {}){
+  return new Promise(async resolve => {
+    await $.wait(2000);
+    $.post(taskUrl(function_id, body), (err, resp, data) => {
+      try {
+        if (err) {
+          console.log('\n种豆得豆: API查询请求失败 ‼️‼️')
+          console.log(`function_id:${function_id}`)
+          $.logErr(err);
+        } else {
+          data = JSON.parse(data);
+        }
+      } catch (e) {
+        $.logErr(e, resp);
+      } finally {
+        resolve(data);
+      }
+    })
+  })
+}
+function taskUrl(function_id, body) {
+  body["version"] = "9.0.0.1";
+  body["monitor_source"] = "plant_app_plant_index";
+  body["monitor_refer"] = "";
   return {
-    // url: `${JD_API_HOST}?functionId=${function_id}&body=${escape(JSON.stringify(body))}&appid=ld&client=apple&clientVersion=&networkType=&osVersion=&uuid=`,
     url: JD_API_HOST,
-    body: `functionId=${function_id}&body=${JSON.stringify(body)}&appid=ld&client=apple&clientVersion=&networkType=&osVersion=&uuid=`,
+    body: `functionId=${function_id}&body=${escape(JSON.stringify(body))}&appid=ld&client=apple&area=5_274_49707_49973&build=167283&clientVersion=9.1.0`,
     headers: {
       'Cookie': cookie,
       'Host': 'api.m.jd.com',
       'Accept': '*/*',
       'Connection': 'keep-alive',
-      'User-Agent': 'JD4iPhone/167249 (iPhone;iOS 13.5.1;Scale/3.00)',
+      'User-Agent': 'JD4iPhone/167249 (iPhone;iOS 13.6.1;Scale/3.00)',
       'Accept-Language': 'zh-Hans-CN;q=1,en-CN;q=0.9',
       'Accept-Encoding': 'gzip, deflate, br',
       'Content-Type': "application/x-www-form-urlencoded"
     }
   }
 }
-
-// function taskurl(function_id, body) {
-//     return {
-//         url: `${JD_API_HOST}?functionId=${function_id}`,
-//         body: body, //escape`functionId=${function_id}&body=${JSON.stringify(body)}&appid=wh5`
-//         headers: {
-//             Cookie: cookie,
-//         },
-//         method: "POST",
-//     }
-// }
-
-function sleep(response) {
-  console.log('休息一下');
-  setTimeout(() => {
-    console.log('休息结束');
-    Task.next(response)
-  }, 2000);
-}
-
 function getParam(url, name) {
-  var reg = new RegExp("(^|&)" + name + "=([^&]*)(&|$)", "i");
-  var r = url.match(reg);
+  const reg = new RegExp("(^|&)" + name + "=([^&]*)(&|$)", "i")
+  const r = url.match(reg)
   if (r != null) return unescape(r[2]);
   return null;
 }
